Test/Bugfix: Make sure that `assert.ok` checks real truth
Previously, we just did `assert.ok`, which tests *truthiness*, but not
actual truth.  This resulted in tests "passing", even when none of the
`id` tests was actually running, and one of the string equality tests
was also not running.  Those two oversights are fixed in this commit,
along with making sure that all of the tests are now checking against
`true`.